requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
jdcal==1.4.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
openpyxl==3.0.4
pyTelegramBotAPI==3.7.2
requests==2.24.0
urllib3==1.25.10
xlrd==1.2.0
